Wrangler and C3: Move the cli folder of C3 into @cloudflare/cli and make Wrangler and C3 depend on it
We want to start using @clack/core and C3-like components in Wrangler, so the logical next step is to make these components available to wrangler and C3.

`./helpers/cli.ts` in C3 is now index.ts in @cloudflare/cli, didn't include openInBrowser,C3_DEFAULTS and WRANGLER_DEFAULTS.
`./helpers/colors` in C3 is now colors.ts in @cloudflare/cli.
`./helpers/interactive.ts` in C3 is now interactive.ts in @cloudflare/cli.
