+++ .changeset/itchy-ducks-listen.md
@@ -1,0 +1,7 @@
+---
+"@cloudflare/cli": major
+"create-cloudflare": patch
+"wrangler": patch
+---
+
+Move helper cli files of C3 into @cloudflare/cli and make Wrangler and C3 depend on it
