+++ packages/cli/.eslintrc.js
@@ -1,0 +1,5 @@
+module.exports = {
+	root: true,
+	extends: ["@cloudflare/eslint-config-worker"],
+	ignorePatterns: ["dist"],
+};
