+++ packages/cli/README.md
@@ -1,0 +1,3 @@
+# Cloudflare CLI
+
+All things related to rendering the CLI for workers-sdk.
