+++ packages/cli/__tests__/cli.test.ts
@@ -1,0 +1,8 @@
+import { describe, expect, test } from "vitest";
+import { space } from "..";
+
+describe("cli", () => {
+	test("test spaces", () => {
+		expect(space(300)).toHaveLength(300);
+	});
+});
