+++ packages/cli/tsconfig.json
@@ -1,0 +1,13 @@
+{
+	"root": true,
+	"extends": "@cloudflare/workers-tsconfig/tsconfig.json",
+	"include": ["**/*.ts"],
+	"exclude": ["node_modules"],
+	"compilerOptions": {
+		"module": "CommonJS",
+		"allowJs": false,
+		"outDir": "dist",
+		"types": ["node"],
+		"resolveJsonModule": true
+	}
+}
