+++ packages/cli/turbo.json
@@ -1,0 +1,6 @@
+{
+	"$schema": "http://turbo.build/schema.json",
+	"extends": ["//"],
+
+	"pipeline": {}
+}
