+++ packages/cli/vite.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+	test: {
+		include: ["**/__tests__/**/*.{test,spec}.{ts,js,tsx,jsx}"],
+		testTimeout: 30000,
+		setupFiles: "./vite.setup.ts",
+	},
+});
