+++ packages/cli/vite.setup.ts
@@ -1,0 +1,3 @@
+import { vi } from "vitest";
+
+vi.mock("log-update");
