+++ packages/create-cloudflare/package.json
@@ -42,8 +42,8 @@
 	"devDependencies": {
 		"@babel/parser": "^7.21.3",
 		"@babel/types": "^7.21.4",
-		"@clack/core": "^0.3.2",
 		"@clack/prompts": "^0.6.3",
+		"@cloudflare/cli": "workspace:*",
 		"@cloudflare/eslint-config-worker": "*",
 		"@cloudflare/workers-tsconfig": "workspace:*",
 		"@cloudflare/workers-types": "^4.20230419.0",
@@ -58,7 +58,6 @@
 		"@types/yargs": "^17.0.22",
 		"@typescript-eslint/eslint-plugin": "^5.55.0",
 		"@typescript-eslint/parser": "^5.55.0",
-		"ansi-escapes": "^6.2.0",
 		"chalk": "^5.2.0",
 		"command-exists": "^1.2.9",
 		"cross-spawn": "^7.0.3",
@@ -67,7 +66,6 @@
 		"execa": "^7.1.1",
 		"glob": "^10.3.3",
 		"haikunator": "^2.1.2",
-		"log-update": "^5.0.1",
 		"open": "^8.4.0",
 		"pnpm": "^8.6.11",
 		"recast": "^0.22.0",
