+++ packages/create-cloudflare/src/frameworks/astro/index.ts
@@ -1,10 +1,10 @@
-import { logRaw } from "helpers/cli";
-import { brandColor, dim } from "helpers/colors";
+import { logRaw } from "@cloudflare/cli";
+import { brandColor, dim } from "@cloudflare/cli/colors";
 import { npmInstall, runCommand, runFrameworkGenerator } from "helpers/command";
 import { compatDateFlag } from "helpers/files";
 import { detectPackageManager } from "helpers/packages";
 import { getFrameworkCli } from "../index";
-import type { PagesGeneratorContext, FrameworkConfig } from "types";
+import type { FrameworkConfig, PagesGeneratorContext } from "types";
 
 const { npx, dlx } = detectPackageManager();
