+++ packages/create-cloudflare/src/frameworks/gatsby/index.ts
@@ -1,9 +1,9 @@
+import { inputPrompt } from "@cloudflare/cli/interactive";
 import { runFrameworkGenerator } from "helpers/command";
 import { compatDateFlag } from "helpers/files";
-import { inputPrompt } from "helpers/interactive";
 import { detectPackageManager } from "helpers/packages";
 import { getFrameworkCli } from "../index";
-import type { PagesGeneratorContext, FrameworkConfig } from "types";
+import type { FrameworkConfig, PagesGeneratorContext } from "types";
 
 const { npm, dlx } = detectPackageManager();
