+++ packages/create-cloudflare/src/frameworks/nuxt/index.ts
@@ -1,10 +1,9 @@
-import { logRaw } from "helpers/cli";
+import { logRaw } from "@cloudflare/cli";
 import { npmInstall, runFrameworkGenerator } from "helpers/command";
-import { compatDateFlag } from "helpers/files";
-import { writeFile } from "helpers/files";
+import { compatDateFlag, writeFile } from "helpers/files";
 import { detectPackageManager } from "helpers/packages";
 import { getFrameworkCli } from "../index";
-import type { PagesGeneratorContext, FrameworkConfig } from "types";
+import type { FrameworkConfig, PagesGeneratorContext } from "types";
 
 const { npm, dlx } = detectPackageManager();
