+++ packages/create-cloudflare/src/frameworks/qwik/index.ts
@@ -1,9 +1,9 @@
-import { endSection } from "helpers/cli";
+import { endSection } from "@cloudflare/cli";
 import { npmInstall, runCommand, runFrameworkGenerator } from "helpers/command";
 import { compatDateFlag } from "helpers/files";
 import { detectPackageManager } from "helpers/packages";
 import { getFrameworkCli } from "../index";
-import type { PagesGeneratorContext, FrameworkConfig } from "types";
+import type { FrameworkConfig, PagesGeneratorContext } from "types";
 
 const { npm, npx, dlx } = detectPackageManager();
