+++ packages/create-cloudflare/src/frameworks/svelte/index.ts
@@ -1,6 +1,6 @@
-import { logRaw, updateStatus } from "helpers/cli";
+import { logRaw, updateStatus } from "@cloudflare/cli";
+import { brandColor, dim, blue } from "@cloudflare/cli/colors";
 import { parseTs, transformFile } from "helpers/codemod";
-import { blue, dim, brandColor } from "helpers/colors";
 import {
 	installPackages,
 	npmInstall,
@@ -11,7 +11,7 @@ import { detectPackageManager } from "helpers/packages";
 import { getFrameworkCli } from "../index";
 import { platformInterface } from "./templates";
 import type * as recast from "recast";
-import type { PagesGeneratorContext, FrameworkConfig } from "types";
+import type { FrameworkConfig, PagesGeneratorContext } from "types";
 
 const { npm, dlx } = detectPackageManager();
