+++ packages/create-cloudflare/src/helpers/args.ts
@@ -1,10 +1,11 @@
+import { crash, logRaw } from "@cloudflare/cli";
+import { getRenderers, inputPrompt } from "@cloudflare/cli/interactive";
 import yargs from "yargs";
 import { hideBin } from "yargs/helpers";
 import { version } from "../../package.json";
 import { templateMap } from "../templateMap";
-import { C3_DEFAULTS, WRANGLER_DEFAULTS, crash, logRaw } from "./cli";
-import { getRenderers, inputPrompt } from "./interactive";
-import type { PromptConfig } from "./interactive";
+import { C3_DEFAULTS, WRANGLER_DEFAULTS } from "./cli";
+import type { PromptConfig } from "@cloudflare/cli/interactive";
 import type { C3Args } from "types";
 
 export const parseArgs = async (argv: string[]): Promise<Partial<C3Args>> => {
